rename all @redux usages to @zus & remove dependencies
