+++ package.json
@@ -10,7 +10,6 @@
     "@types/node": "^14.0.13",
     "@types/react": "^16.9.38",
     "@types/react-dom": "^16.9.8",
-    "@types/react-redux": "^7.1.9",
     "@types/styled-jsx": "^2.2.8",
     "@types/three": "^0.103.2",
     "bootstrap": "^4.5.0",
@@ -24,12 +23,8 @@
     "react-dom": "^16.13.1",
     "react-dropzone": "^11.2.4",
     "react-helmet": "^6.1.0",
-    "react-redux": "^7.2.0",
     "react-scripts": "3.4.1",
     "react-three-fiber": "^5.3.1",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-thunk": "^2.3.0",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^1.1.0",
     "styled-jsx": "^3.3.0",
