+++ tsconfig.paths.json
@@ -7,7 +7,7 @@
       "@constants/*": ["src/constants/*"],
       "@libs/*": ["src/libs/*"],
       "@pages/*": ["src/pages/*"],
-      "@redux/*": ["src/redux/*"],
+      "@zus/*": ["src/zustand/*"],
       "@utils/*": ["src/utils/*"]
     }
   }
